// src/app/models/Release/Release.ts
export interface Release {
    IDRELEASESTATE: number;
    IDRELEASETYPE: number;
    IDSECTION: number;
    TITLERELEASE: string;
    FIRSTSTATEMENTRELEASE: string;
    CONTENTRELEASE: string;
    SUBSECTION: string;
    MAINIMAGE: string;
    SINGLETITLE: string;
    SINGLEIMG: String;
}